wip: add replies

+++ src/services/api/reviews/remove-like.ts
@@ -1,11 +1,17 @@
+import { LikeReviewValues } from '@/hooks/use-reviews/use-reviews.types'
 import { supabase } from '@/services/supabase'
 
-export const removeLikeService = async (id: number) => {
+export const removeLikeService = async (values: LikeReviewValues) => {
+  const { reviewId, userId } = values
+
   const { error, data } = await supabase
-    .from('review_likes')
+    .from('likes')
     .delete()
-    .eq('id', id)
+    .eq('entity_type', 'REVIEW')
+    .eq('review_id', reviewId)
+    .eq('user_id', userId)
 
   if (error) throw new Error(error.message)
+
   return data
 }
